XRDDEV-754: Refactor .properties-files to yaml-format
