+++ src/proxy-ui-api/src/main/java/org/niis/xroad/restapi/config/IpThrottlingFilter.java
@@ -70,15 +70,15 @@ public class IpThrottlingFilter extends GenericFilterBean {
     LoadingCache<String, Bucket> bucketCache = CacheBuilder.newBuilder()
             .expireAfterAccess(EXPIRE_BUCKETS_FROM_CACHE_MINUTES, TimeUnit.MINUTES)
             .build(new CacheLoader<String, Bucket>() {
-                       public Bucket load(String key) {
-                           Bucket bucket = createStandardBucket();
-                           return bucket;
-                       }
-                   }
-            );
+                public Bucket load(String key) {
+                    Bucket bucket = createStandardBucket();
+                    return bucket;
+                }
+            });
 
     /**
      * create a new bucket
+     *
      * @return
      */
     private Bucket createStandardBucket() {
