+++ src/proxy-ui-api/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.context.ApplicationListener=org.niis.xroad.restapi.config.StartStopListener
-org.springframework.boot.env.EnvironmentPostProcessor=org.niis.xroad.restapi.config.DatabasePropertiesEnvironmentPostProcessor,org.niis.xroad.restapi.config.SslPropertiesEnvironmentPostProcessor
+org.springframework.boot.env.EnvironmentPostProcessor=org.niis.xroad.restapi.config.DatabasePropertiesEnvironmentPostProcessor,org.niis.xroad.restapi.config.SslPropertiesEnvironmentPostProcessor,org.niis.xroad.restapi.config.CommonPropertyEnvironmentPostProcessor
