--- src/proxy-ui-api/src/main/resources/application-development.properties
@@ -1,5 +1,0 @@
-# non-harmful development profile settings
-# these can be included in the production jar
-
-logging.level.org.springframework.security=DEBUG
-web.security.debug=true
