+++ src/proxy-ui-api/src/main/resources/application-development.yml
@@ -1,0 +1,12 @@
+# non-harmful development profile settings
+# these can be included in the production jar
+
+logging:
+  level:
+    org:
+      springframework:
+        security: DEBUG
+
+web:
+  security:
+    debug: true
