--- src/proxy-ui-api/src/main/resources/application.properties
@@ -1,21 +1,0 @@
-# shared properties for dev/prod runtime
-# test runtime has a separate application.properties
-
-spring.jpa.hibernate.ddl-auto=none
-spring.jpa.hibernate.show-sql=true
-
-spring.jpa.properties.hibernate.dialect = ee.ria.xroad.common.db.CustomPostgreSQLDialect
-spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
-
-# TLS
-# (can be overridden with external ssl.properties)
-server.ssl.key-store=/etc/xroad/ssl/proxy-ui-api.p12
-server.ssl.key-store-password=proxy-ui-api
-server.ssl.enabled=true
-server.ssl.ciphers=TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256,TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384,TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256,TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384
-server.ssl.protocol=TLS
-server.ssl.enabled-protocols=TLSv1.2
-
-logging.config=file:/etc/xroad/conf.d/proxy-ui-api-logback.xml
-
-spring.profiles.include=nontest
