+++ src/proxy-ui-api/src/main/resources/application.yml
@@ -1,4 +1,32 @@
-# define file upload endpoints in yaml since it is more convenient than with .properties
+# shared properties for dev/prod runtime
+# test runtime has a separate application.properties
+spring:
+  jpa:
+    hibernate:
+      ddl-auto: none
+      show-sql: true
+      naming:
+        physical-strategy: org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
+    properties:
+      hibernate:
+        dialect: ee.ria.xroad.common.db.CustomPostgreSQLDialect
+  profiles:
+    include: nontest
+
+# TLS
+# (can be overridden with external ssl.properties)
+server:
+  ssl:
+    key-store: /etc/xroad/ssl/proxy-ui-api.p12
+    key-store-password: proxy-ui-api
+    enabled: true
+    ciphers: TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256,TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384,TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256,TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384
+    protocol: TLS
+    enabled-protocols: TLSv1.2
+
+logging:
+  config: file:/etc/xroad/conf.d/proxy-ui-api-logback.xml
+
 file-upload-endpoints:
   endpointDefinitions:
     -
