--- src/proxy-ui-api/src/main/resources/common-application.properties
@@ -1,35 +1,0 @@
-## properties that are not specific to a runtime (test or prod)
-
-spring.thymeleaf.cache: false
-
-# default is 30m
-# server.servlet.session.timeout=1m
-server.port=4000
-
-spring.mvc.throw-exception-if-no-handler-found=true
-spring.resources.add-mappings=false
-#server.error.include-stacktrace=always
-#server.error.include-exception=true
-
-## *.hbm.xml files are loaded from classpath
-spring.jpa.mapping-resources=identifiers.hbm.xml,serverconf.hbm.xml,apikey.hbm.xml
-spring.jpa.hibernate.ddl-auto=none
-spring.jpa.hibernate.show-sql=true
-
-server.servlet.session.cookie.secure=true
-
-# max-http-post-size applies only to multipart/form-data requests (although name is misleading)
-# https://github.com/spring-projects/spring-boot/issues/18521
-# we do not use multipart/form-data, so we can se the limit to a strict value
-# if container is changed to undertow, these settings will apply to ALL requests
-# due to https://github.com/spring-projects/spring-boot/issues/18555
-server.tomcat.max-http-post-size=1KB
-spring.servlet.multipart.max-file-size=1KB
-spring.servlet.multipart.max-request-size=1KB
-
-spring.jackson.default-property-inclusion=NON_NULL
-
-request.sizelimit.regular=50KB
-request.sizelimit.binary.upload=10MB
-ratelimit.requests.per.second=20
-ratelimit.requests.per.minute=600
