--- src/proxy-ui-api/src/test/resources/application.properties
@@ -1,16 +1,0 @@
-# properties for test runtime
-
-spring.jpa.hibernate.ddl-auto=create-drop
-spring.jpa.generate-ddl=true
-spring.jpa.show-sql=true
-
-logging.level.org.springframework.security=DEBUG
-
-##server.error.include-stacktrace=always
-##server.error.include-exception=true
-
-logging.level.org.springframework.transaction.interceptor=TRACE
-log4j.logger.org.springframework.orm.jpa=DEBUG
-log4j.logger.org.springframework.transaction=DEBUG
-
-spring.profiles.active=test
