+++ src/proxy-ui-api/src/test/resources/application.yml
@@ -1,0 +1,31 @@
+# properties for test runtime
+
+spring:
+  jpa:
+    hibernate:
+      ddl-auto: create-drop
+    generate-ddl: true
+    show-sql: true
+  profiles:
+    active: test
+
+logging:
+  level:
+    org:
+      springframework:
+        security: DEBUG
+        transaction:
+          interceptor: TRACE
+
+log4j:
+  logger:
+    org:
+      springframework:
+        orm:
+          jpa: DEBUG
+        transaction: DEBUG
+
+#server:
+#  error:
+#    include-stacktrace: always
+#    include-exception: true
